Move RSS generation out of layout
Cleaner this way, and maybe faster
