+++ src/app/layout.jsx
@@ -2,9 +2,6 @@ import AppBar from '@/components/appBar/AppBar';
 import ThemeRegistry from '@/components/themeRegistry/ThemeRegistry';
 import getInitColorSchemeScript from '@mui/system/cssVars/getInitColorSchemeScript';
 import { Box } from '@mui/joy';
-import RSS from 'rss';
-import { writeFileSync } from 'fs';
-import { getSortedPostMetadata } from '@/posts';
 
 const isDev = process.env.NODE_ENV !== 'production';
 const rootUrl = isDev ? 'http://localhost:3000' : 'https://sixfold-origami.com';
@@ -26,8 +23,6 @@ export const metadata = {
 };
 
 export default function RootLayout({ children }) {
-	generateRSS();
-
 	return (
 		<html lang="en" data-joy-color-scheme="dark">
 			<body>
@@ -50,30 +45,4 @@ export default function RootLayout({ children }) {
 			</body>
 		</html>
 	);
-}
-
-// I tried putting this in a prebuild script but the imports made it a pain
-function generateRSS() {
-	const feedOptions = {
-		title: `${title} | RSS Feed`,
-		description,
-		site_url: rootUrl,
-		feed_url: `${rootUrl}/rss.xml`,
-		image_url: `${rootUrl}/opengraph-image.png`,
-		pubDate: new Date(),
-		copyright: `All rights reserved ${new Date().getFullYear()}`,
-	};
-	const feed = new RSS(feedOptions);
-
-	const posts = getSortedPostMetadata();
-	posts.map((p) => {
-		feed.item({
-			title: p.title,
-			description: p.subtitle,
-			url: `${rootUrl}/articles/${p.slug}`,
-			date: p.date,
-		});
-	});
-
-	writeFileSync('./public/rss.xml', feed.xml({ indent: true }));
 }
